tests for pointToSubstitute
